+++ src/test/kotlin/com/oriolsoler/costcontroler/integration/acceptance/ShowAllCostsFeature.kt
@@ -9,14 +9,42 @@ import org.springframework.security.test.web.servlet.request.SecurityMockMvcRequ
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.model
+import java.math.BigDecimal
 
 
 abstract class ShowAllCostsFeature : IntegrationTest() {
     @Test
     fun `should show all user costs`() {
-        val cost1 = registerCost("Description cost 1", "Category1", "Subcategory1", "Comment1", 1.0, "Oriol", false, null)
-        val cost2 = registerCost("Description cost 2", "Category2", "Subcategory2", "Comment2", 2.0, "Oriol", false, null)
-        val cost3 = registerCost("Description cost 3", "Category3", "Subcategory3", "Comment3", 3.0, "Jonny", false, null)
+        val cost1 = registerCost(
+            "Description cost 1",
+            "Category1",
+            "Subcategory1",
+            "Comment1",
+            BigDecimal.valueOf(1.0),
+            "Oriol",
+            false,
+            null
+        )
+        val cost2 = registerCost(
+            "Description cost 2",
+            "Category2",
+            "Subcategory2",
+            "Comment2",
+            BigDecimal.valueOf(2.0),
+            "Oriol",
+            false,
+            null
+        )
+        val cost3 = registerCost(
+            "Description cost 3",
+            "Category3",
+            "Subcategory3",
+            "Comment3",
+            BigDecimal.valueOf(3.0),
+            "Jonny",
+            false,
+            null
+        )
 
         mvc.perform(get("/show").with(user("Oriol")))
             .andExpect(model().attribute("costs", hasSize<Int>(2)))
